+++ packages/react/types/Active.d.ts
@@ -3,12 +3,12 @@ import React from "react";
 import { ReactElement } from "react";
 import { CuriRouter, Response } from "@curi/core";
 import { HickoryLocation } from "@hickory/root";
+export declare type ActiveChildren = (active: boolean) => ReactElement<any>;
 export interface ActiveProps {
-    children: ReactElement<any>;
+    children: ActiveChildren;
     name: string;
     params?: object;
     partial?: boolean;
-    merge(props: object): object;
     extra?(l: HickoryLocation, d: object): boolean;
     details?: object;
 }